refatored page objet
